feat(plugin-table): ship the reference table UI as `plugin-table/ui`
The table chrome lived in the playground: ~1800 lines of geometry,
interaction, and render code that any consumer wanting the playground's
table experience would have to copy. It consumes only public primitives
(`useEditor`, traversal, the plugin's custom events), so it moves into
the package behind a `./ui` subpath export, keeping the root entry
headless.

`table-chrome.tsx`, `table-cell-style.ts`, `table-metrics.ts`,
`table-drag.ts`, `table-reorder.ts`, and `table-overflow.ts` move
verbatim; the `TableContainer`/`TableRow`/`TableCell` renders extract
out of the playground's `plugin.table.tsx`, which shrinks to the
container composition and its playground-owned cell content renders
(list items, images, callouts), the seam consumers customize. The four
lucide icons are inlined as plain SVGs so the entry ships without an
icon dependency; they are the defaults a consumer-provided `icons`
object will later override.

Theming is CSS custom properties: the playground's `--color-table-*`
tokens become the public `--pt-plugin-table-*` contract (bare `--pt-*`
stays reserved for core editor surface). Light defaults and the
structural rules ship in `plugin-table/ui/styles.css` under
`:where(:root)`: token defaults must lose to any consumer override
regardless of stylesheet load order, and plain `:root` ties with a
consumer's `.dark` at specificity 0-1-0, letting bundle order decide.
The playground now consumes the subpath like any consumer: it imports
the stylesheet and overrides the tokens under `.dark`.

Net behavior unchanged: light and dark renders, column selection
visuals, and the table menu are screenshot-verified against the
pre-extraction chrome; all 106 plugin tests pass unmodified.
