+++ apps/playground/src/editor.css
@@ -199,49 +199,3 @@
 [data-level='10'] {
   padding-left: calc(var(--list-padding) * 9);
 }
-
-/**
- * Table layout only. Cell borders, header base, selection outline/overlay and
- * corner radii are inline per cell (see `table-cell-style.ts`), matching the
- * design prototype.
- */
-.playground-table-chrome {
-  position: relative;
-  margin: 20px 0;
-}
-
-/* Horizontal scroll under a wide table: lighter thumb, gap from the table edge. */
-.table-scroll {
-  scrollbar-width: thin;
-  scrollbar-color: var(--color-table-scrollbar) transparent;
-}
-.table-scroll::-webkit-scrollbar {
-  height: 6px;
-}
-.table-scroll::-webkit-scrollbar-track {
-  margin-top: 8px;
-  background: transparent;
-}
-.table-scroll::-webkit-scrollbar-thumb {
-  background: var(--color-table-scrollbar);
-  border-radius: 3px;
-}
-.table-scroll::-webkit-scrollbar-thumb:hover {
-  background: var(--color-table-scrollbar-hover);
-}
-
-.playground-table {
-  width: 100%;
-  border-collapse: separate;
-  border-spacing: 0;
-  table-layout: fixed;
-}
-
-/* While a multi-cell range is active, the blue rectangle overlay stands in for
-   the selection, so hide the native text-selection highlight. */
-.playground-table[data-cell-range] ::selection {
-  background: transparent;
-}
-.playground-table[data-cell-range] ::-moz-selection {
-  background: transparent;
-}
