+++ packages/plugin-table/src/ui/icons.tsx
@@ -1,0 +1,75 @@
+import type {JSX} from 'react'
+
+/**
+ * Inlined from lucide (ISC), so the UI entry ships without an icon
+ * dependency. These are the built-in defaults; a consumer-provided `icons`
+ * object is the planned override point.
+ */
+
+type IconProps = {
+  size: number
+}
+
+export function EllipsisIcon(props: IconProps): JSX.Element {
+  return (
+    <Icon {...props}>
+      <circle cx="12" cy="12" r="1" />
+      <circle cx="19" cy="12" r="1" />
+      <circle cx="5" cy="12" r="1" />
+    </Icon>
+  )
+}
+
+export function PanelTopIcon(props: IconProps): JSX.Element {
+  return (
+    <Icon {...props}>
+      <rect width="18" height="18" x="3" y="3" rx="2" />
+      <path d="M3 9h18" />
+    </Icon>
+  )
+}
+
+export function TableIcon(props: IconProps): JSX.Element {
+  return (
+    <Icon {...props}>
+      <path d="M12 3v18" />
+      <rect width="18" height="18" x="3" y="3" rx="2" />
+      <path d="M3 9h18" />
+      <path d="M3 15h18" />
+    </Icon>
+  )
+}
+
+export function Trash2Icon(props: IconProps): JSX.Element {
+  return (
+    <Icon {...props}>
+      <path d="M10 11v6" />
+      <path d="M14 11v6" />
+      <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6" />
+      <path d="M3 6h18" />
+      <path d="M8 6V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" />
+    </Icon>
+  )
+}
+
+function Icon({
+  size,
+  children,
+}: IconProps & {children: React.ReactNode}): JSX.Element {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={2}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+    >
+      {children}
+    </svg>
+  )
+}
