+++ packages/plugin-table/src/ui/index.ts
@@ -1,0 +1,1 @@
+export {TableCell, TableContainer, TableRow} from './table-render'
