+++ packages/plugin-table/src/ui/table-metrics.ts
